+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Thin item dividers for all lists
 - Confirmation dialogs for destructive actions (logout and item deletion)
 - Confirmation dialog for discard item and item authorization changes action
+- Password generator dialog on item screen
 
 ## Changed
 - Allow to copy item details of read-only items (do not disable them anymore) - the delete button is hidden instead of disabled to avoid user confusion (like the save accept toolbar icon)